requirements.txt
openai
SpeechRecognition
python-docx
PyAudio
python-dotenv
